Added in new "Social" class for UserProfiles (to contain followers, following and inbox)
